+++ CHANGELOG.md
@@ -11,7 +11,8 @@ of [keepachangelog.com][2].
 
 ### Added
 
-- None
+- [#22](https://github.com/jaredbeck/minitest_to_rspec/pull/22) -
+  add `--newline` option to preserve newline
 
 ### Fixed